This is a combination of 7 commits.
Side nav drawer improvements; split component from MainLayout component; improved animation integrity; nav link active status classes

Side nav auto-close behavior; custom breakpoint hook

Disable open-drawer button on breakpoint

Disable open-drawer button on breakpoint

Remove dated import

Tiny alignment change for close-drawer button

chore: format files from GitHub Action
