+++ client/components/layout/MainLayout.tsx
@@ -1,119 +1,28 @@
-import React, { useState } from 'react';
-import { Outlet, Link as RouterLink } from 'react-router-dom';
-// import Footer from './Footer';
-// import LinktaLogo from './LinktaLogoWithText';
-// import TopNavigationBar from './TopNavigationBar';
-// import useDynamicNavigation from '@/client/hooks/useDynamicNavigation';
-import {
-  Box,
-  Button,
-  ButtonGroup,
-  Drawer,
-  Link,
-  Typography,
-} from '@mui/material';
-import {
-  ArrowDropDown,
-  ChevronLeftOutlined,
-  ChevronRightOutlined,
-  AccountCircleOutlined,
-  AddCircleOutline,
-  HelpOutlineOutlined,
-  SettingsOutlined,
-} from '@mui/icons-material';
+import React, { useEffect, useState } from 'react';
+import { Outlet } from 'react-router-dom';
+import { Box, Button, ButtonGroup, Typography } from '@mui/material';
+import { ArrowDropDown } from '@mui/icons-material';
+import SideNavDrawer from '@client/components/common/SideNavDrawer';
+import useMatchMedia from '@client/hooks/useMediaMatch';
 import '@client/styles/MainLayout.css';
 
 /**
- * MainLayout manages the app's global layout, using `useDynamicNavigation` for route-based UI adjustments. It presents a consistent header featuring the LinktaLogo, with the top navigation bar and footer rendered conditionally as dictated by the current route's needs.
- *
- * - `showTopNavBar` and `showFooter`: Boolean values from `useDynamicNavigation` determine the visibility of the TopNavigationBar and Footer, respectively.
- * - The `Outlet` component handles rendering of route-specific content in the main section.
+ * MainLayout provides the app's global UI layout and the Router Outlet.
+ * It presents a consistent header featuring the LinktaLogo, with the side navigation bar rendered as open or compact conditionally based on screen width.
+ * The `Outlet` component handles rendering of route-specific content in the main section.
  */
 const MainLayout: React.FC = () => {
-  // const { showTopNavBar, showFooter } = useDynamicNavigation();
-  const [open, setOpen] = useState(true);
-  const toggleDrawer = (newOpen: boolean) => () => {
-    setOpen(newOpen);
-  };
-  const mountedStyle = { animation: 'opacity-in 300ms ease-in' };
-  const unmountedStyle = {
-    animation: 'opacity-out 200ms ease-in',
-    animationFillMode: 'forwards',
+  const breakpoint = 768;
+  const matching = useMatchMedia(breakpoint);
+
+  const [drawerOpen, setDrawerOpen] = useState(!matching);
+  const toggleDrawer = () => {
+    !matching && setDrawerOpen(!drawerOpen);
   };
 
-  const DrawerList = (
-    <Box className="side-nav-bar">
-      <Box className="linkta-logo-drawer">
-        <img
-          className="linkta-drawer-image"
-          src="../assets/linkta-logo-transparent.svg"
-          alt="The Linkta.io logo at approximately 50 pixels square, located in the upper-left corner of the page."
-        />
-      </Box>
-
-      <Link
-        className="side-nav-link"
-        underline="none"
-        mt={3}
-        component={RouterLink}
-        to="/sign-in"
-      >
-        <AccountCircleOutlined />
-        <Typography variant="caption">test.user@linkta.org</Typography>
-      </Link>
-
-      <Link
-        className="side-nav-link"
-        underline="none"
-        component={RouterLink}
-        to="/generate"
-      >
-        <AddCircleOutline />
-        <Typography variant="caption">Explore a New Topic</Typography>
-      </Link>
-
-      <Box
-        className="recent-user-inputs"
-        mt={5}
-        pl={2}
-      >
-        <Typography
-          variant="body2"
-          color={'primary.contrastText'}
-        >
-          Recent
-        </Typography>
-      </Box>
-
-      <Link
-        className="side-nav-link"
-        underline="none"
-        mt={'auto'}
-        component={RouterLink}
-        to="/help-and-feedback"
-      >
-        <HelpOutlineOutlined />
-        <Typography variant="caption">Help and Feedback</Typography>
-      </Link>
-
-      <Link
-        className="side-nav-link"
-        underline="none"
-        component={RouterLink}
-        to="/settings"
-      >
-        <SettingsOutlined />
-        <Typography variant="caption">Settings</Typography>
-      </Link>
-
-      <Button
-        className="drawer-close-button"
-        onClick={toggleDrawer(false)}
-        startIcon={<ChevronLeftOutlined />}
-        sx={{ marginBottom: '20px', paddingInline: '20px' }}
-      ></Button>
-    </Box>
-  );
+  useEffect(() => {
+    matching && setDrawerOpen(false);
+  }, [ matching ]);
 
   return (
     <>
@@ -125,21 +34,18 @@ const MainLayout: React.FC = () => {
 
       {/* this Box contains the Linkta main layout components - top bar, side nav, router outlet */}
       <Box
-        className={`static-layout ${open ? 'layout-open' : 'layout-closed'}`}
+        className={`static-layout ${drawerOpen ? 'layout-open' : 'layout-closed'}`}
         component="main"
       >
-        {!open ? (
-          <Box className="linkta-logo-container">
-            <img
-              className="linkta-logo-image"
-              src="../assets/linkta-logo-transparent.svg"
-              alt="The Linkta.io logo at approximately 50 pixels square, located in the upper-left corner of the page."
-              style={open ? unmountedStyle : mountedStyle}
-            />
-          </Box>
-        ) : (
-          <Box className="drawer-open-logo-ph"></Box>
-        )}
+        <Box
+          className={`linkta-logo-container ${drawerOpen ? 'layout-open' : 'layout-closed'}`}
+        >
+          <img
+            className={`linkta-logo-image ${drawerOpen ? 'layout-open' : 'layout-closed'}`}
+            src="../assets/linkta-logo-web.png"
+            alt="The Linkta.io logo at approximately 50 pixels square, located in the upper-left corner of the page."
+          />
+        </Box>
 
         <Box className="top-nav-bar">
           <ButtonGroup
@@ -173,71 +79,15 @@ const MainLayout: React.FC = () => {
           </ButtonGroup>
         </Box>
 
-        {!open ? (
-          <Box
-            className="side-nav-mini"
-            pb={2}
-          >
-            <Link
-              className="side-mini-button"
-              mt={3}
-              style={open ? unmountedStyle : mountedStyle}
-              component={RouterLink}
-              to="/sign-in"
-            >
-              <AccountCircleOutlined />
-            </Link>
-
-            <Link
-              className="side-mini-button"
-              style={open ? unmountedStyle : mountedStyle}
-              component={RouterLink}
-              to="/generate"
-            >
-              <AddCircleOutline />
-            </Link>
-
-            <Link
-              className="side-mini-button"
-              sx={{ marginTop: 'auto' }}
-              style={open ? unmountedStyle : mountedStyle}
-              component={RouterLink}
-              to="/help-and-feedback"
-            >
-              <HelpOutlineOutlined />
-            </Link>
-
-            <Link
-              className="side-mini-button"
-              style={open ? unmountedStyle : mountedStyle}
-              component={RouterLink}
-              to="/settings"
-            >
-              <SettingsOutlined />
-            </Link>
-
-            <Button
-              className="drawer-open-button"
-              onClick={toggleDrawer(true)}
-            >
-              <ChevronRightOutlined />
-            </Button>
-          </Box>
-        ) : (
-          <Box></Box>
-        )}
+        <SideNavDrawer
+          drawerOpen={drawerOpen}
+          matching={matching}
+          toggleDrawer={toggleDrawer}
+        />
 
         <Box className="router-outlet">
           <Outlet />
         </Box>
-
-        <Drawer
-          open={open}
-          onClose={toggleDrawer(false)}
-          variant="persistent"
-        >
-          {DrawerList}
-        </Drawer>
       </Box>
     </>
   );
