+++ client/components/prompt-input/PromptInputForm.tsx
@@ -1,0 +1,8 @@
+import { Box } from '@mui/material';
+import React from 'react';
+
+const PromptInputForm = () => {
+  return <Box>PromptInputForm Placeholder</Box>;
+};
+
+export default PromptInputForm;
