+++ client/components/prompt-input/PromptInputInstructions.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const PromptInputInstructions = () => {
+  return <div>PromptInputInstructions Placeholder</div>;
+};
+
+export default PromptInputInstructions;
