--- client/features/prompt-input-page/components/PromptInputInstructions.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const PromptInputInstructions = () => {
-  return <div>PromptInputInstructions Placeholder</div>;
-};
-
-export default PromptInputInstructions;
