+++ client/styles/SideNavDrawer.css
@@ -1,0 +1,114 @@
+:root {
+  --secondary-main: #ffa51b;
+}
+
+.side-nav-compact {
+  opacity: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
+  width: 60px;
+  padding-bottom: 20px;
+  align-items: center;
+  box-sizing: border-box;
+  -moz-box-sizing: border-box;
+  -webkit-box-sizing: border-box;
+}
+
+.side-nav-bar {
+  display: flex;
+  flex-direction: column;
+  width: 186px;
+  gap: 5px;
+  padding-bottom: 20px;
+}
+
+button.first-button {
+  margin-top: 80px;
+}
+
+.side-nav-link > *,
+.side-mini-button > *,
+.drawer-open-button > svg,
+.drawer-close-button > svg {
+  color: #f5f5f5;
+}
+
+.drawer-open-button.disabled > svg {
+  color: rgba(245, 245, 245, 0.4);
+}
+
+.side-nav-link,
+.side-mini-button {
+  cursor: pointer;
+  transition: 200ms ease-in;
+}
+
+.side-nav-link:hover,
+.side-mini-button:hover {
+  box-shadow: inset 0px 0px 4px rgba(159, 133, 18, 0.4);
+  background-color: rgba(255, 255, 255, 0.15);
+  box-sizing: border-box;
+  -moz-box-sizing: border-box;
+  -webkit-box-sizing: border-box;
+}
+
+.side-nav-link.active > *,
+.side-mini-button.active > * {
+  color: var(--secondary-main);
+}
+
+.side-nav-link {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+  padding: 7px 0px 7px 14px;
+  -webkit-user-select: none; /* Safari */
+  -ms-user-select: none; /* IE 10 and IE 11 */
+  user-select: none; /* Standard syntax */
+}
+
+.side-mini-button {
+  padding: 5px 14px;
+}
+
+.side-nav-button > .MuiTypography-root {
+  margin-right: auto;
+  font-size: 11px;
+}
+
+button.drawer-close-button {
+  justify-content: start;
+}
+
+div.MuiDrawer-paper {
+  background-color: transparent;
+  flex-direction: row;
+  width: 186px;
+}
+
+@keyframes opacity-in {
+  0% {
+    opacity: 0;
+    visibility: hidden;
+  }
+  50% {
+    opacity: 0;
+    visibility: hidden;
+  }
+  100% {
+    opacity: 1;
+    visibility: visible;
+  }
+}
+
+@keyframes opacity-out {
+  0% {
+    opacity: 1;
+    visibility: visible;
+  }
+  100% {
+    opacity: 0;
+    visibility: hidden;
+  }
+}
